A complete experiment containing one or more models

An S4 class to represent a self-enclosed modeling experiment or intentioned set of models It contains a list of `roleModel`objects, a list of `roleParams` to use for each model, a data.frame summary of the models, and some author metadata

Usage

roleExperiment(allParams)

Arguments

allParams

the list of model params to use for each model

Value

a ready-to-run `roleExperiment`

Slots

modelRuns

a list of `roleModel`objects

allParams

a list of `roleParams` to use for each model

authorMeta

a named string vector that keep track of author metadata It contains values for "author", "date", "description", "info", where each element is named by its respective string When the model is saved with `writeRole` a text file is generated using this metadata

experimentMeta

data.frame of summarizing metadata for all the models of the experiment (Need to chat with Andy about the exact intentions of this before writing more here)

Examples